actions refctoring 2 ver
