+++ news/views/news.py
@@ -1,10 +1,10 @@
-from mixins import NewsMixin
+from mixins.like import LikedMixin
+from mixins.news import NewsMixin
 from rest_framework import filters, viewsets
 from rest_framework.exceptions import ValidationError
 
 from ..models.news import News
 from ..serializers.news import NewsSerializer
-from news.mixins import LikedMixin
 from news.permission import AuthorOrReadOnlyNews
 from users.models.user import User
