+++ users/mixins/author.py
@@ -4,10 +4,9 @@ from rest_framework.decorators import action
 from rest_framework.request import HttpRequest
 from rest_framework.response import Response
 
-from .models.author import Author
-from .serializers.profile import ProfileSerializer
+from ..models.author import Author
 from news.serializers.news import NewsSerializer
-from users.permission import IsModerator, UserOwnerOrReadOnly
+from users.permission import IsModerator
 
 
 class AuthorMixin:
@@ -50,28 +49,3 @@ class AuthorMixin:
         }
 
         return Response(data)
-
-
-class UserMixin:
-    @action(
-        methods=['GET'],
-        url_path='profile',
-        detail=False,
-        permission_classes=[UserOwnerOrReadOnly, ],
-    )
-    def profile(self, request: HttpRequest):
-        serializer = ProfileSerializer(request.user)
-        return Response(serializer.data)
-
-    @profile.mapping.patch
-    def update_my_model(self, request: HttpRequest):
-        serializer = ProfileSerializer(request.user, data=request.data)
-        serializer.is_valid(raise_exception=True)
-        self.perform_update(serializer)
-        return Response(serializer.data)
-
-    @profile.mapping.delete
-    def profile_delete(self, request: HttpRequest):
-        serializer = ProfileSerializer(request.user)
-        serializer.destroy()
-        return Response({'detail': 'Success'}, status=204)
