+++ users/mixins/user.py
@@ -1,0 +1,31 @@
+from rest_framework.decorators import action
+from rest_framework.request import HttpRequest
+from rest_framework.response import Response
+
+from ..serializers.profile import ProfileSerializer
+from users.permission import UserOwnerOrReadOnly
+
+
+class UserMixin:
+    @action(
+        methods=['GET'],
+        url_path='profile',
+        detail=False,
+        permission_classes=[UserOwnerOrReadOnly, ],
+    )
+    def profile(self, request: HttpRequest):
+        serializer = ProfileSerializer(request.user)
+        return Response(serializer.data)
+
+    @profile.mapping.patch
+    def update_my_model(self, request: HttpRequest):
+        serializer = ProfileSerializer(request.user, data=request.data)
+        serializer.is_valid(raise_exception=True)
+        self.perform_update(serializer)
+        return Response(serializer.data)
+
+    @profile.mapping.delete
+    def profile_delete(self, request: HttpRequest):
+        serializer = ProfileSerializer(request.user)
+        serializer.destroy()
+        return Response({'detail': 'Success'}, status=204)
